fix(local-display): 消除 lvgl 层的平台宽度写死（Windows 侧曾红 13 处）
拉取 fe9810a 后，Windows 上 `cargo check -p local-display --all-targets` 报 13 处
E0308/E0277（生产 3 + 测试 11），而 Linux 侧正常 —— 根因是 **C 枚举的签名宽度随平台
浮动**：bindgen 把 `lv_dir_t` / `lv_scrollbar_mode_t` / `lv_part_t` / `lv_state_t` /
`lv_label_long_mode_t` 映射为 **MSVC `c_int`(i32) / Linux `c_uint`(u32)**，而代码里
写死了 `as u32`（当初正是为了让 Linux 编过）⇒ 换平台即红。

改法：**去掉写死宽度、统一用 bindgen 别名**（一份代码两平台，不引入 cfg 分叉）：
- `widgets.rs` 的 `scroll_dir` / `scrollbar_mode`：删掉多余的 `as u32` —— FFI 返回值
  与 `Dir` / `ScrollMode` 包裹的别名本就同型；
- `style.rs::to_sys`：两侧各自 cast 到 `lv_style_selector_t`（C 侧 `typedef uint32_t`
  ⇒ **恒为 u32**），与 LVGL 的 `(uint32_t)part | (uint32_t)state` 同构；
- 测试 3 处：同口径（对系统别名两侧同型比较，不再单侧 cast）。

⚠️ **Linux 侧不受影响（按构造成立，不是"应该没事"）**：被删的 cast 在 Linux 上是
空操作（`u32 as u32`），新增的 cast 同样是空操作 ⇒ Linux 侧生成代码逐字不变。
**本机无法验 Linux**（无 Linux C 工具链，build.rs 要给 aarch64 编 LVGL）⇒ 以 CI 的
`hmi`(aarch64) 与 `lint`/`test`(x86_64 Linux) 为准。

验证（Windows）：`check --workspace --all-targets` 0 error；`test -p local-display`
= 369 + 6 + 6 + 3 全绿（与基线一致）；`lvgl/**` 内对 bindgen 别名的宽度 cast 已清零。

+++ mupc/crates/local-display/src/lvgl/tests_a3.rs
@@ -131,7 +131,8 @@ pub(super) fn widgets_chain() {
     value.set_size(40, 20);
     value.set_long_mode(LongMode::WRAP);
     assert_eq!(value.text().as_deref(), Some("128"));
-    assert_eq!(LongMode::WRAP.raw(), sys::LV_LABEL_LONG_MODE_WRAP as u32);
+    // 两侧同为 `lv_label_long_mode_t`（平台浮动别名）⇒ **不得**在一侧写 `as u32`（MSVC 下是 i32）。
+    assert_eq!(LongMode::WRAP.raw(), sys::LV_LABEL_LONG_MODE_WRAP);
     // 状态机制（TT-10 按钮禁用 / §6.2 步进器越界共用这一入口）。
     assert!(!widgets::has_state(&bare, State::DISABLED));
     widgets::set_state(&bare, State::DISABLED, true);
@@ -355,8 +356,10 @@ pub(super) fn widgets_chain() {
     // ── Important (d)：`Style::set_width` + `Part::{KNOB, SELECTED}` ─────────
     {
         // 取值与 C 端 `lv_part_t` 逐位一致（评审已核，测试再钉一遍）。
-        assert_eq!(Part::KNOB.raw(), sys::LV_PART_KNOB as u32);
-        assert_eq!(Part::SELECTED.raw(), sys::LV_PART_SELECTED as u32);
+        // 两侧同为 `lv_part_t`（平台浮动别名）⇒ **不得**在一侧写 `as u32`（MSVC 下是 i32）；
+        // 下面两行与裸字面量比较，字面量按左侧类型推断，故两平台都成立。
+        assert_eq!(Part::KNOB.raw(), sys::LV_PART_KNOB);
+        assert_eq!(Part::SELECTED.raw(), sys::LV_PART_SELECTED);
         assert_eq!(Part::KNOB.raw(), 0x030000, "LV_PART_KNOB");
         assert_eq!(Part::SELECTED.raw(), 0x040000, "LV_PART_SELECTED");
         // 已登记的四个取值与原值不冲突（无重复 / 错值）。
